+++ templates/skeleton/app/routes.ts
@@ -1,7 +1,5 @@
 import {flatRoutes} from '@remix-run/fs-routes';
-import {layout, type RouteConfig} from '@remix-run/route-config';
+import {type RouteConfig} from '@remix-run/route-config';
 import {hydrogenRoutes} from '@shopify/hydrogen';
 
-export default hydrogenRoutes([
-  layout('./layout.tsx', (await flatRoutes())),
-]) satisfies RouteConfig;
+export default hydrogenRoutes(await flatRoutes()) satisfies RouteConfig;
